+++ application/views/itemadd.php
@@ -1,0 +1,1 @@
+<a href="/mtceFleet/add"><input type="button" value="Add a new item"/></a>