+++ application/views/itemlistFlight.php
@@ -1,0 +1,13 @@
+{pagination}
+<table class="table">
+ 	<tr>
+		<th>FlightID</th>
+		<th>VehicleID</th>
+		<th>DepartureTime</th>
+		<th>ArrivalTime</th>
+		<th>Base</th>
+		<th>Dest</th>
+	</tr>
+		
+        {display_flight}    
+</table>