+++ application/views/itemnav.php
@@ -1,0 +1,6 @@
+<ul class="nav nav-pills">
+        <li><a href="/mtceFleet/page/{first}">First</a></li>
+        <li><a href="/mtceFleet/page/{previous}">Previous</a></li>
+        <li><a href="/mtceFleet/page/{next}">Next</a></li>
+        <li><a href="/mtceFleet/page/{last}">Last</a></li>
+</ul>