+++ application/views/itemnavFlight.php
@@ -1,0 +1,6 @@
+<ul class="nav nav-pills">
+        <li><a href="/mtceFlight/page/{first}">First</a></li>
+        <li><a href="/mtceFlight/page/{previous}">Previous</a></li>
+        <li><a href="/mtceFlight/page/{next}">Next</a></li>
+        <li><a href="/mtceFlight/page/{last}">Last</a></li>
+</ul>