+++ application/views/oneitemFlightx.php
@@ -1,0 +1,8 @@
+<tr>
+	<td><a href="/mtceFlight/edit/{flightID}"><input type="button" value="{flightID}"/></a></td>
+	<td>{vehicleID}</td>
+	<td>{departureTime}</td>
+	<td>{arrivalTime}</td>
+	<td>{base}</td>
+	<td>{dest}</td>
+</tr>