+++ application/views/oneitemx.php
@@ -1,0 +1,9 @@
+<tr>
+	<td><a href="/mtceFleet/edit/{vehicleID}"><input type="button" value="{vehicleID}"/></a></td>
+	<td>{model}</td>
+	<td>{seats}</td>
+	<td>{reach}</td>
+	<td>{cruise}</td>
+	<td>{takeoff}</td>
+	<td>{hourly}</td>
+</tr>